+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Features
 * Created SEO Tools as a standalone Django application.
 * Provided ability to post HTML in the comments.
 * Added Celery integration support.
+* Contact form submission now uses Celery queues.
 
 
 Improvements